+++ lib/roseflow/openai.rb
@@ -1,6 +1,12 @@
 # frozen_string_literal: true
 
+require "roseflow/types"
+
 require_relative "openai/version"
+require_relative "openai/client"
+require_relative "openai/config"
+require_relative "openai/operation_handler"
+require_relative "openai/provider"
 
 module Roseflow
   module OpenAI