+++ lib/roseflow/openai/chat_message.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require "roseflow/chat/message"
+
+module Types
+  module OpenAI
+    FunctionCallObject = Types::Hash
+    StringOrObject = Types::String | FunctionCallObject
+    StringOrArray = Types::String | Types::Array
+  end
+end
+
+module Roseflow
+  module OpenAI
+    class ChatMessage < Roseflow::Chat::Message
+      attribute? :function_call, Types::OpenAI::FunctionCallObject
+    end
+  end
+end