+++ spec/roseflow/operation_handler_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Roseflow
+  module OpenAI
+    RSpec.describe OperationHandler do
+      let(:valid_operations) { described_class::VALID_OPERATIONS }
+    end
+  end
+end