+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.2.0] - 2023-07-24
+
+- Roseflow::OpenAI::Model now implements Roseflow::AI::Model API spec.
+- API operations are now composed and validated before calling the API.
+
 ## [0.1.0] - 2023-04-26
 
 - Initial release